requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.30
aiomysql==0.2.0
